test: merge all quote sources tests
